requirements.txt
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
pgvector==0.2.4
requests==2.31.0
python-dotenv==1.0.1
streamlit==1.31.1
numpy==1.26.4
together==0.2.5
scikit-learn==1.4.0  # For anomaly detection 